model/requirements.txt
beautifulsoup4==4.11.1
redis
jupyter==1.0.0
matplotlib==3.5.2
seaborn==0.12.1
nltk==3.7
pandas==1.4.2
scikit-learn==1.1.1
typer==0.4.0
click==8.0.4
spacy
unidecode==1.3.4
gensim==4.2.0
pytest==7.1.1
pyyaml==6.0
scipy==1.8.0
plotly_express==0.4.0
protobuf==3.20.*

transformers
tensorflow==2.8.0
